refactor: unify test crate
