+++ tests/model/mod.rs
@@ -1,0 +1,13 @@
+mod button;
+mod footer;
+mod image;
+mod legal;
+mod vstack;
+
+pub use button::Button;
+pub use footer::Footer;
+pub use image::Image;
+pub use legal::Legal;
+pub use vstack::VStack;
+
+pub trait View {}
