--- view_test/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "view_test"
-version = "0.1.0"
-authors = ["richard <richard>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-view = {path="../view"}
